refactor: separate CLI output and share provider requests
